% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_resolve_pkg_path.R
\name{check_if_binary}
\alias{check_if_binary}
\title{Throw an error if the package is a pre-built binary}
\usage{
check_if_binary(pkg_dir)
}
\arguments{
\item{pkg_dir}{Path to the package directory.}
}
\description{
Throw an error if the package is a pre-built binary
}
\keyword{internal}
