% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_combine_vignettes.R
\name{combine_vignettes}
\alias{combine_vignettes}
\title{Helper function to combine package vignettes}
\usage{
combine_vignettes(pkg_path)
}
\arguments{
\item{pkg_path}{Path to the package source directory.}
}
\value{
A single string containing the combined vignettes from the package.
}
\description{
Helper function to combine package vignettes
}
\keyword{internal}
