% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_resolve_remote.R
\name{find_pkg_dir}
\alias{find_pkg_dir}
\title{Find Package Directory Within Extracted Bundle}
\usage{
find_pkg_dir(extract_dir, subdir = NULL)
}
\arguments{
\item{extract_dir}{Directory where bundle was extracted}

\item{subdir}{Optional subdirectory path}
}
\value{
Path to package directory
}
\description{
Find Package Directory Within Extracted Bundle
}
\keyword{internal}
