% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_resolve_remote.R
\name{get_remote_cache_dir}
\alias{get_remote_cache_dir}
\title{Get Cache Directory for Remote Package}
\usage{
get_remote_cache_dir(parsed, cache_path)
}
\arguments{
\item{parsed}{Parsed reference}

\item{cache_path}{Base cache path}
}
\value{
Path to cache directory
}
\description{
Get Cache Directory for Remote Package
}
\keyword{internal}
