% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_resolve_pkg_path.R
\name{is_binary_pkg}
\alias{is_binary_pkg}
\title{Check if a directory contains a pre-built binary package}
\usage{
is_binary_pkg(pkg_dir)
}
\arguments{
\item{pkg_dir}{Path to the package directory.}
}
\value{
Logical indicating if it is a binary package.
}
\description{
Check if a directory contains a pre-built binary package
}
\keyword{internal}
