% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_resolve_remote.R
\name{is_remote_reference}
\alias{is_remote_reference}
\title{Check if String is a Remote Package Reference}
\usage{
is_remote_reference(pkg)
}
\arguments{
\item{pkg}{Character string to check}
}
\value{
TRUE if it looks like a remote reference
}
\description{
Check if String is a Remote Package Reference
}
\keyword{internal}
