% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_resolve_remote.R
\name{parse_remote_url}
\alias{parse_remote_url}
\title{Parse GitHub or GitLab Web URL}
\usage{
parse_remote_url(url)
}
\arguments{
\item{url}{The full URL string}
}
\value{
List with components: type, user, repo, ref, subdir
}
\description{
Parse GitHub or GitLab Web URL
}
\keyword{internal}
