% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_resolve_remote.R
\name{parse_remote_ref}
\alias{parse_remote_ref}
\title{Parse Remote Reference String}
\usage{
parse_remote_ref(ref)
}
\arguments{
\item{ref}{Character string reference}
}
\value{
List with components: type, user, repo, ref, subdir
}
\description{
Supports formats:
\itemize{
\item "user/repo" -> GitHub shorthand
\item "github::user/repo" -> Explicit GitHub
\item "gitlab::user/repo" -> GitLab
\item "bitbucket::user/repo" -> Bitbucket
\item "user/repo@ref" -> With commit/branch/tag
\item "user/repo/subdir" -> With subdirectory
\item "user/repo/subdir@ref" -> Combined
\item "https://github.com/user/repo/tree/ref/subdir" -> Web URL
}
}
\keyword{internal}
