% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_code.R
\name{rdd_extract_code}
\alias{rdd_extract_code}
\title{Extract R Source Code from a Package}
\usage{
rdd_extract_code(
  pkg,
  file = NULL,
  include_tests = FALSE,
  force_fetch = FALSE,
  cache_path = getOption("rdocdump.cache_path")
)
}
\arguments{
\item{pkg}{A \code{character} string specifying the package. This can be:
\itemize{
\item an installed package name,
\item a full path to a package source directory,
\item a full path to a package archive file (tar.gz), or
\item a package name not installed (which will then be downloaded from CRAN).
}}

\item{file}{Optional. Save path for the output text file. If set, the function will return the path to the file instead of the combined text. Defaults to \code{NULL}.}

\item{include_tests}{\code{logical}. If \code{TRUE}, for non-installed packages, the function will also include R source code from the \code{tests} directory. Defaults to \code{FALSE}.}

\item{force_fetch}{\code{logical}. If \code{TRUE}, the package source will be fetched from CRAN even if the package is installed locally. Default is \code{FALSE}.}

\item{cache_path}{A \code{character} string specifying the directory to use as a cache. Defaults to the value of \code{getOption("rdocdump.cache_path")}.}
}
\value{
A single string containing the combined R source code from the package.
}
\description{
This function extracts the R source code from a package. For installed packages, it retrieves the
package namespace and deparses all functions found in the package. For package source directories
or archives (non-installed packages), it reads all \code{.R} files from the \code{R} directory and, optionally, from the \code{tests}
directory.
}
\examples{
# Extract only R source code from an installed package.
code <- rdd_extract_code("stats")
cat(code)

# Extract R source code from a package source directory (and optionally include test files).
code_with_tests <- rdd_extract_code("mypackage", include_tests = TRUE)
cat(code_with_tests)
}
