% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_code.R
\name{rdd_extract_code}
\alias{rdd_extract_code}
\title{Extract R Source Code from a Package}
\usage{
rdd_extract_code(
  pkg,
  file = NULL,
  include_tests = FALSE,
  include_roxygen = FALSE,
  force_fetch = FALSE,
  cache_path = getOption("rdocdump.cache_path")
)
}
\arguments{
\item{pkg}{A \code{character} string specifying the package. This can be:
\itemize{
\item an installed package name,
\item a full path to a package source directory,
\item a full path to a package archive file (tar.gz), or
\item a package name not installed (which will then be downloaded from CRAN).
}}

\item{file}{Optional. Save path for the output text file. If set, the function will return the path to the file instead of the combined text. Defaults to \code{NULL}.}

\item{include_tests}{\code{logical}. If \code{TRUE}, for non-installed packages, the function will also include R source code from the \code{tests} directory. Defaults to \code{FALSE}.}

\item{include_roxygen}{\code{logical}. If \code{TRUE}, roxygen2 documentation lines (lines starting with "#'") from R files will be included in the output. Defaults to \code{FALSE}.}

\item{force_fetch}{\code{logical}. If \code{TRUE}, the package source will be fetched from CRAN even if the package is installed locally. Default is \code{FALSE}.}

\item{cache_path}{A \code{character} string specifying the directory to use as a cache. Defaults to the value of \code{getOption("rdocdump.cache_path")}.}
}
\value{
A single string containing the combined R source code (and, optionally, roxygen2 documentation) from the package.
}
\description{
This function extracts the R source code from a package. For installed packages, it retrieves the
package namespace and deparses all functions found in the package. For package source directories
or archives (non-installed packages), it reads all \code{.R} files from the \code{R} directory and, optionally,
from the \code{tests} directory. Optionally, it can include roxygen2 documentation from these files.
}
\examples{
# Extract only R source code (excluding roxygen2 documentation) from an installed package.
code <- rdd_extract_code("stats")
cat(substr(code, 1, 1000))

# Extract R source code including roxygen2 documentation from a package source directory.
local({
 old_repos <- getOption("repos")
 options(repos = c(CRAN = "https://cran.r-project.org"))
 code_with_roxygen <- rdd_extract_code(
  "ini",
  include_roxygen = TRUE,
  force_fetch = TRUE)
 cat(substr(code_with_roxygen, 1, 1000))
 options(repos = old_repos)
})

# Extract R source code from a package source directory,
# including test files but excluding roxygen2 docs.
local({
 old_repos <- getOption("repos")
 options(repos = c(CRAN = "https://cran.r-project.org"))
 code_with_tests <- rdd_extract_code(
  "ini",
  include_roxygen = TRUE,
  include_tests = TRUE
  force_fetch = TRUE)
 cat(substr(code_with_tests, 1, 1000))
 options(repos = old_repos)
})

}
