% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-or-set-repos.R
\name{rdd_get_repos}
\alias{rdd_get_repos}
\title{Get Current rdocdump Repository Options}
\usage{
rdd_get_repos()
}
\value{
A character vector of repository URLs.
}
\description{
This function returns the current repository URLs used by rdocdump.
}
