% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_cache_path.R
\name{rdd_set_cache_path}
\alias{rdd_set_cache_path}
\title{Set RDocDump Cache Path for the Session}
\usage{
rdd_set_cache_path(path)
}
\arguments{
\item{path}{A character string specifying the directory to be used as the cache path.}
}
\value{
Invisibly returns the new cache path.
}
\description{
This function sets the cache path used to store temporary files (downloaded tar.gz archives
and/or extracted directories) for the current session. The cache path is stored in the option
\code{"rdocdump.cache_path"}.
}
