% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-or-set-repos.R
\name{rdd_set_repos}
\alias{rdd_set_repos}
\title{Set rdocdump Repository Options}
\usage{
rdd_set_repos(repos)
}
\arguments{
\item{repos}{A character vector of repository URLs.}
}
\value{
Invisibly returns the new repository URLs.
}
\description{
This function sets the repository URLs used by rdocdump when fetching package sources.
}
