% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-or-set-repos.R
\name{rdd_set_repos}
\alias{rdd_set_repos}
\title{Set \code{rdocdump} Repository Options}
\usage{
rdd_set_repos(repos)
}
\arguments{
\item{repos}{A character vector of repository URLs.}
}
\value{
Invisibly returns the new repository URLs.
}
\description{
This function sets the package repository URLs used by \code{rdocdump} when fetching package sources. May be useful for setting custom repositories or mirrors. This does not affect the repositories used by \code{install.packages()} in your current R session and/or project.
}
\examples{
# Set rdocdump repository options
rdd_set_repos(c("CRAN" = "https://cloud.r-project.org"))

}
