% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_txt.R
\name{rdd_to_txt}
\alias{rdd_to_txt}
\title{Create Combined Package Documentation as Text}
\usage{
rdd_to_txt(
  pkg,
  file = NULL,
  content = "all",
  force_fetch = FALSE,
  keep_files = "none",
  cache_path = getOption("rdocdump.cache_path")
)
}
\arguments{
\item{pkg}{A \code{character} string specifying the package. This can be:
\itemize{
\item an installed package name,
\item a full path to a package source directory,
\item a full path to a package archive file (tar.gz), or
\item a package name not installed (which will then be downloaded from CRAN).
}}

\item{file}{Optional. Save path for the output text file. If set, the function will return the path to the file instead of the combined text. Defaults to \code{NULL}.}

\item{content}{A character vector specifying which components to include in the output.
Possible values are:
\itemize{
\item \code{"all"}: Include Rd documentation, vignettes, and R source code (default).
\item \code{"docs"}: Include only the Rd documentation.
\item \code{"vignettes"}: Include only the vignettes.
\item \code{"code"}: Include only the R source code.
}
You can specify multiple options (e.g., \code{c("docs", "code")} to include both documentation and source code).}

\item{force_fetch}{\code{logical}. If \code{TRUE}, the package source will be fetched from CRAN as a tar.gz archive even if the package is already installed locally. Default is \code{FALSE}.}

\item{keep_files}{A \code{character} value controlling whether temporary files should be kept.
Possible values are:
\itemize{
\item \code{"none"}: Delete both the tar.gz archive and the extracted files (default).
\item \code{"tgz"}: Keep only the tar.gz archive.
\item \code{"extracted"}: Keep only the extracted files.
\item \code{"both"}: Keep both the tar.gz archive and the extracted files.
}}

\item{cache_path}{A \code{character} string specifying the directory where kept temporary files will be stored.
By default, it uses the value of \code{getOption("rdocdump.cache_path")} which sets the cache directory to the temporary directory of the current R session.}
}
\value{
A single string containing the combined package documentation, vignettes, and/or code as specified by the \code{content} argument.
If the \code{file} argument is set, returns the path to the file.
}
\description{
This function produces a single text output for a package by processing its documentation (Rd files from the package source
or the documentation from already installed packages), vignettes, and/or R source code.
}
\examples{
# Extract all components (documentation, vignettes, and code) from the built-in 'stats' package.
all_components <- rdd_to_txt("stats")
cat(all_components)

# Extract only Rd documentation.
docs_only <- rdd_to_txt("stats", content = "docs")
cat(docs_only)

# Extract both documentation and R source code.
docs_and_code <- rdd_to_txt("stats", content = c("docs", "code"))
cat(docs_and_code)
}
