% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_txt.R
\name{rdd_to_txt}
\alias{rdd_to_txt}
\title{Create Combined Package Documentation as Text}
\usage{
rdd_to_txt(
  pkg,
  file = NULL,
  keep_files = "none",
  cache_path = getOption("rdocdump.cache_path"),
  force_fetch = FALSE
)
}
\arguments{
\item{pkg}{A \code{character} string specifying the package. This can be:
\itemize{
\item an installed package name,
\item a full path to a package source directory,
\item a full path to a package archive file (tar.gz), or
\item a package name not installed (which will then be downloaded from CRAN).
}}

\item{file}{Optional. Save path for the output text file. If set, the function will return back the path to the file instead of the combined text.}

\item{keep_files}{A \code{character} value controlling whether temporary files should be kept.
Possible values are:
\itemize{
\item \code{"none"}: Delete both the tar.gz archive and the extracted files (default).
\item \code{"tgz"}: Keep only the tar.gz archive.
\item \code{"extracted"}: Keep only the extracted files.
\item \code{"both"}: Keep both the tar.gz archive and the extracted files.
}}

\item{cache_path}{A \code{character} string specifying the directory where kept temporary files will be stored. By default, it uses the value of option("rdocdump.cache_path") which sets the cache directory to the temporary directory of the current R session.}

\item{force_fetch}{\code{logical}. If \code{TRUE}, the package source will be fetched from CRAN as a tar.gz archive even if the package is already installed locally Default is \code{FALSE}.}
}
\value{
A single string containing the combined package documentation. If \code{file} argument is set, returns path to the file.
}
\description{
This function produces a single text output for a package by processing its documentation (Rd files from package source, or the documentation from alteady installed packages) and vignettes (from \verb{*.Rmd}, \verb{*.qmd}, \verb{*.md}, \code{and}*.Rnw` files). The user can provide the function with an installed package name, a package source directory, a tar.gz archive, or a package name to download from CRAN.
}
\examples{
# extract documentation for built-in `stats` package
docs <- rdd_to_txt("stats")

\donttest{
# extract documentation for rJavaEnv by downloading its source from CRAN
local({
  old_repos <- getOption("repos")
  options(repos = c(CRAN = "https://cran.r-project.org"))
  docs <- rdd_to_txt("rJavaEnv", force_fetch = TRUE)
  lines <- unlist(strsplit(docs, "\n"))
  # Print the first 3 lines
  cat(head(lines, 3), sep = "\n")
  # Print the last 3 lines
  cat(tail(lines, 3), sep = "\n")
  options(repos = old_repos)
  #unlink(getOption("rdocdump.cache_path"), recursive = TRUE)
})
}
}
