% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_txt.R
\name{rdd_to_txt}
\alias{rdd_to_txt}
\title{Create Combined Package Documentation as Text}
\usage{
rdd_to_txt(
  pkg,
  file = NULL,
  keep_files = "none",
  cache_path = getOption("rdocdump.cache_path"),
  force_fetch = FALSE
)
}
\arguments{
\item{pkg}{A character string specifying the package. This can be:
\itemize{
\item an installed package name,
\item a full path to a package source directory,
\item a full path to a package archive file (tar.gz), or
\item a package name not installed (which will then be downloaded from CRAN).
}}

\item{file}{Optional. Full path for the output text file.}

\item{keep_files}{Character value controlling whether temporary files should be kept.
Possible values are:
\itemize{
\item \code{"none"}: Delete both the tar.gz archive and the extracted files (default).
\item \code{"tgz"}: Keep only the tar.gz archive.
\item \code{"extracted"}: Keep only the extracted files.
\item \code{"both"}: Keep both the tar.gz archive and the extracted files.
}}

\item{cache_path}{Character string specifying the directory where kept temporary files
will be stored. By default, it uses the value of option("rdocdump.cache_path").}

\item{force_fetch}{Logical. If TRUE, the package source will be fetched from CRAN
as a tar.gz archive even if the package is already installed locally.
Default is FALSE.}
}
\value{
A single string containing the combined package documentation.
}
\description{
This function produces a single text output for a package by processing its Rd
files. The package can be provided as an installed package name, a package source
directory, a tar.gz archive, or a package name to download from CRAN. (In the future,
vignettes will be processed as well.)
}
